cleaning up server code, drying up routes
