+++ server/api/tagged.js
@@ -1,19 +1,15 @@
 const router = require("express").Router();
 const Sticker = require("../db/sticker");
 const User = require("../db/user");
-const Friends = require("../db/friends");
 const Sequelize = require("sequelize");
+const getFriendIds = require("./utils/getFriendIds");
 
 router.get("/ByFriends/:userId", async (req, res, next) => {
   try {
     const userId = req.params.userId;
     const user = await User.findByPk(req.params.userId);
-    const friends = await Friends.findAll({
-      where: {
-        userId,
-      },
-    });
-    const friendIds = friends.map((elem) => elem.friendId);
+    const friendIds = await getFriendIds(userId);
+
     //gets all friends stickers where the user is tagged
     const taggedStickers = await Sticker.findAll({
       where: {
@@ -23,7 +19,6 @@ router.get("/ByFriends/:userId", async (req, res, next) => {
         },
       },
     });
-
     res.json(taggedStickers);
   } catch (err) {
     next(err);
@@ -34,12 +29,7 @@ router.get("/ByRandos/:userId", async (req, res, next) => {
   try {
     const userId = req.params.userId;
     const user = await User.findByPk(userId);
-    const friends = await Friends.findAll({
-      where: {
-        userId,
-      },
-    });
-    const friendIds = friends.map((elem) => elem.friendId);
+    const friendIds = await getFriendIds(userId);
     //gets all friends stickers where the user is tagged
     const taggedStickers = await Sticker.findAll({
       where: {
@@ -51,7 +41,6 @@ router.get("/ByRandos/:userId", async (req, res, next) => {
         },
       },
     });
-
     res.json(taggedStickers);
   } catch (err) {
     next(err);
