+++ server/api/users.js
@@ -1,6 +1,7 @@
 const router = require("express").Router();
 const User = require("../db/user");
 const Sequelize = require("sequelize");
+const pickPropsFromObj = require("./utils/pickPropsFromObj");
 
 //get all users; GET /users
 router.get("/", function (req, res, next) {
@@ -9,13 +10,9 @@ router.get("/", function (req, res, next) {
 
 router.post("/friendSearch", async function (req, res, next) {
   const query = req.body.query.toLowerCase();
-  console.log("req.body.query: ", query);
   const potentialFriends = await User.findAll({
     attributes: ["userName", "id"],
     where: {
-      // userName: {
-      //   [Sequelize.Op.like]: `%${query}%`,
-      // },
       userName: Sequelize.where(
         Sequelize.fn("LOWER", Sequelize.col("userName")),
         "LIKE",
@@ -25,18 +22,6 @@ router.post("/friendSearch", async function (req, res, next) {
   });
   res.json(potentialFriends);
 });
-
-//picks the array of props from an object and doesn't assign keys to the returned object
-//that are falsey
-const pickPropsFromObj = (props, obj) => {
-  return props.reduce((acc, prop) => {
-    const value = obj[prop];
-    if (obj[prop]) {
-      return { ...acc, [prop]: value };
-    }
-    return acc;
-  }, {});
-};
 
 //create user; POST /users
 router.post("/", async (req, res, next) => {
@@ -50,10 +35,8 @@ router.post("/", async (req, res, next) => {
   }
 });
 
-//create user; POST /users
-
+//TO BE DEVELOPED
 //get single user; GET /users/:id
-
 //update user; PUT /users/:id
 //delete user; DELETE /users/:id
