+++ server/api/utils/pickPropsFromObj.js
@@ -1,0 +1,11 @@
+const pickPropsFromObj = (props, obj) => {
+  return props.reduce((acc, prop) => {
+    const value = obj[prop];
+    if (obj[prop]) {
+      return { ...acc, [prop]: value };
+    }
+    return acc;
+  }, {});
+};
+
+module.exports = pickPropsFromObj;
